requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
pyarrow
apache-airflow
pyyaml
psycopg2-binary
apache-airflow-providers-celery
db-dtypes
google-cloud-bigquery-storage
python-dotenv

